+++ CHANGELOG.md
@@ -614,6 +614,7 @@
 * Moved the explicit routing function to an advanced routing API
 * Added saga data snapshotting + snapshot storages for SQL Server and MongoDB
 * Updated Newtonsoft JSON.NET dep to 7.0.1
+* Added file system-based saga snapshot storage
 
 
 [AndreaCuneo]: https://github.com/AndreaCuneo